Fixed BoundingBox class, DataManager and removed Voxel, ToolPath and its writer

+++ Model/Point3D.cpp
@@ -2,16 +2,16 @@
 #include "Point3D.h"
 
 Point3D::Point3D(): mx(0),my(0),mz(0) {}
-Point3D::Point3D(const float& inX, const float& inY, const float& inZ) :mx(inX), my(inY), mz(inZ) {}
+Point3D::Point3D(const double& inX, const double& inY, const double& inZ) :mx(inX), my(inY), mz(inZ) {}
 Point3D::~Point3D() {}
 
-void Point3D::setX(const float& inX) {mx = inX;}
-void Point3D::setY(const float& inY) {my = inY;}
-void Point3D::setZ(const float& inZ) {mz = inZ;}
+void Point3D::setX(const double& inX) {mx = inX;}
+void Point3D::setY(const double& inY) {my = inY;}
+void Point3D::setZ(const double& inZ) {mz = inZ;}
 
-const float& Point3D::X() const{return mx;}
-const float& Point3D::Y() const{return my;}
-const float& Point3D::Z() const{return mz;}
+const double& Point3D::X() const{return mx;}
+const double& Point3D::Y() const{return my;}
+const double& Point3D::Z() const{return mz;}
 
 bool Point3D::operator<(const Point3D& other) const
 {
